Set recording delay to 0 (disable it) by default
Since the app requires optional permissions for showing the
countdown, it makes sense to disable it out of the box. So, when
the user will enable this feature, he'll know that if he doesn't
grant the overlay permissions he will not see the countdown overlay
and hence understand that the recording is being postponed.